moving functions into qa.py
